pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tracking-analysis"
version = "0.1.0"
description = "Psychophysics tracking data analysis tool for blob tracking experiments"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Psychophysics Research Group"}
]
keywords = ["psychophysics", "tracking", "analysis", "visualization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "matplotlib>=3.4.0",
    "plotly>=5.0.0",
    "streamlit>=1.20.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0.0",
    "jinja2>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.10.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
stats = [
    "statsmodels>=0.13.0",
    "pingouin>=0.5.0",
]

[project.scripts]
tracking-analysis = "tracking_analysis.cli.main:main"

[project.urls]
Repository = "https://github.com/your-org/tracking-analysis"

[tool.setuptools.packages.find]
where = ["."]
include = ["tracking_analysis*"]

[tool.pytest.ini_options]
testpaths = ["tracking_analysis/tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
